Update index.pb
Use the Expo Application Services (EAS) to build an .apk (Android Package) file for your project.

EAS is a relatively new offering from Expo, and the most recently added tool in its toolchain.

The .apk file generated using EAS can be used to install your app as a standalone app, one that does not require the Expo Go client, to your emulator or mobile device. It is not intended to be used to submit to the Android Play Store or Apple App Store.

Note: We will not ask you to create builds for submitting to the Android Play Store or Apple App Store in this course, because both require payment of a registration fee to submit apps.

If you are interested in submitting apps to the Play Store or App Store, you can find out how to do so from the Expo documentation linked in the Additional Resources.